+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/logging/LogMessageKeys.java
@@ -240,6 +240,7 @@ public enum LogMessageKeys {
     // query plan
     PLAN,
     PLAN_HASH,
+    QUERY,
 
     // error
     ERROR,