+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/LuceneIndexQueryPlan.java
@@ -25,18 +25,29 @@ import com.apple.foundationdb.record.ExecuteProperties;
 import com.apple.foundationdb.record.IndexEntry;
 import com.apple.foundationdb.record.IndexScanType;
 import com.apple.foundationdb.record.PlanHashable;
+import com.apple.foundationdb.record.RecordCoreException;
 import com.apple.foundationdb.record.RecordCursor;
 import com.apple.foundationdb.record.RecordMetaData;
 import com.apple.foundationdb.record.TupleRange;
+import com.apple.foundationdb.record.logging.LogMessageKeys;
+import com.apple.foundationdb.record.metadata.Index;
+import com.apple.foundationdb.record.metadata.RecordType;
+import com.apple.foundationdb.record.metadata.expressions.GroupingKeyExpression;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreBase;
 import com.apple.foundationdb.record.query.expressions.Comparisons;
+import com.apple.foundationdb.record.query.plan.AvailableFields;
+import com.apple.foundationdb.record.query.plan.IndexKeyValueToPartialRecord;
 import com.apple.foundationdb.record.query.plan.ScanComparisons;
+import com.apple.foundationdb.record.query.plan.plans.QueryPlanUtils;
+import com.apple.foundationdb.record.query.plan.plans.QueryResult;
 import com.apple.foundationdb.record.query.plan.plans.RecordQueryIndexPlan;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.protobuf.Message;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.Collection;
 import java.util.Objects;
 import java.util.concurrent.ExecutorService;
 
@@ -138,6 +149,62 @@ public class LuceneIndexQueryPlan extends RecordQueryIndexPlan {
         return plan;
     }
 
+    /**
+     * Override here to have specific logic to build the {@link QueryResult} for lucene auto complete suggestion result.
+     */
+    @Nonnull
+    @Override
+    public <M extends Message> RecordCursor<QueryResult> executePlan(@Nonnull FDBRecordStoreBase<M> store,
+                                                                      @Nonnull EvaluationContext context,
+                                                                      @Nullable byte[] continuation,
+                                                                      @Nonnull ExecuteProperties executeProperties) {
+        final RecordMetaData metaData = store.getRecordMetaData();
+        final Index index = metaData.getIndex(indexName);
+        final Collection<RecordType> recordTypes = metaData.recordTypesForIndex(index);
+        if (recordTypes.size() != 1) {
+            throw new RecordCoreException("No lucene index should span multiple record types")
+                    .addLogInfo(LogMessageKeys.INDEX_NAME, indexName);
+        }
+
+        if (scanType == IndexScanType.BY_LUCENE_AUTO_COMPLETE) {
+            final RecordType recordType = recordTypes.iterator().next();
+            final RecordCursor<IndexEntry> entryRecordCursor = executeEntries(store, context, continuation, executeProperties);
+            return entryRecordCursor
+                    .map(QueryPlanUtils.getCoveringIndexEntryToPartialRecordFunction(store, recordType.getName(), indexName, getToPartialRecord(index, recordType), scanType))
+                    .map(QueryResult::of);
+        } else {
+            return super.executePlan(store, context, continuation, executeProperties);
+        }
+    }
+
+    /**
+     * Get the {@link IndexKeyValueToPartialRecord} instance for an {@link IndexEntry} representing a result of Lucene auto-complete suggestion.
+     * The partial record contains the suggestion in the field where it is indexed from, and the grouping keys if there are any.
+     */
+    @VisibleForTesting
+    public static IndexKeyValueToPartialRecord getToPartialRecord(@Nonnull Index index, @Nonnull RecordType recordType) {
+        final IndexKeyValueToPartialRecord.Builder builder = IndexKeyValueToPartialRecord.newBuilder(recordType);
+
+        KeyExpression root = index.getRootExpression();
+        if (root instanceof GroupingKeyExpression) {
+            KeyExpression groupingKey = ((GroupingKeyExpression) root).getGroupingSubKey();
+            for (int i = 0; i < groupingKey.getColumnSize(); i++) {
+                AvailableFields.addCoveringField(groupingKey, AvailableFields.FieldData.of(IndexKeyValueToPartialRecord.TupleSource.KEY, i), builder);
+            }
+        }
+
+        builder.addRequiredMessageFields();
+        if (!builder.isValid(true)) {
+            throw new RecordCoreException("Missing required field for auto complete result record")
+                    .addLogInfo(LogMessageKeys.INDEX_NAME, index.getName())
+                    .addLogInfo(LogMessageKeys.RECORD_TYPE, recordType.getName());
+        }
+
+        builder.addRegularCopier(new LuceneIndexKeyValueToPartialRecordUtils.LuceneAutoCompleteCopier());
+
+        return builder.build();
+    }
+
     @Nonnull
     @Override
     public <M extends Message> RecordCursor<IndexEntry> executeEntries(@Nonnull final FDBRecordStoreBase<M> store,
@@ -147,8 +214,7 @@ public class LuceneIndexQueryPlan extends RecordQueryIndexPlan {
         final TupleRange range = groupingComparisons == null ? comparisons.toTupleRange() : comparisons.append(groupingComparisons).toTupleRange(store, context);
         final RecordMetaData metaData = store.getRecordMetaData();
         LuceneScanProperties scanProperties = new LuceneScanProperties(executeProperties, sort, executorService, reverse);
-        RecordCursor<IndexEntry> indexEntryRecordCursor = store.scanIndex(metaData.getIndex(indexName), scanType, range,
+        return store.scanIndex(metaData.getIndex(indexName), scanType, range,
                 continuation, scanProperties);
-        return indexEntryRecordCursor;
     }
 }